21.05.21

+++ lib/models_and_data/Class/models_and_data.dart
@@ -12,7 +12,7 @@ class Resturant {
   String photoResturant;
   String logoResturant;
   String photoInfoPage;
-  double ratingResturant;
+  int ratingResturant;
   double deliveryPrice;
   double minimumOrder;
   int deliveryDuration;
